CRUD idea

+++ src/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { IdeaDTO } from './idea.dto';
@@ -19,16 +19,33 @@ export class IdeaService {
   }
 
   async read(id: string){
-    return await this.ideaRepository.findOne({ where: { id } });
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if(!idea){
+      throw new HttpException('Not found',  HttpStatus.NOT_FOUND);
+    }
+    return idea;
   }
 
   async update(id: string, data: Partial<IdeaDTO>){
-   await this.ideaRepository.update( { id }, data); 
-   return await this.ideaRepository.findOne({ id });
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if(!idea) {
+      throw new HttpException('Not found',  HttpStatus.NOT_FOUND);
+    }
+    const upIdea = await this.ideaRepository.update({ id }, data);
+    if(!upIdea) {
+      throw new HttpException('Not found',  HttpStatus.NOT_FOUND);
+    }
+    const newIdea = await this.ideaRepository.findOne({ where: { id } });
+    return newIdea;
   }
 
   async destroy(id: string){
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if(!idea) {
+      throw new HttpException('Not found',  HttpStatus.NOT_FOUND);
+    }
     await this.ideaRepository.delete({ id});
-    return { deleted: true };
+    // return { deleted: true };
+    return idea;
   }
 }
